feat: implement EmployeeVideoPlayer component for enhanced video playback experience
